+++ core/src/de/tum/cit/ase/bomberquest/map/Chest.java
@@ -90,13 +90,13 @@ public class Chest implements Drawable {
     }
     
     @Override
-    public float getX() {
-        return x;
+    public int getX() {
+        return (int) x;
     }
     
     @Override
-    public float getY() {
-        return y;
+    public int getY() {
+        return (int) y;
     }
 
     public boolean ChestOpened() {
